refactor: correct typo package name
rename pacage workanalysis to wordanalysis
move RedisHashCacheService class'es package
