+++ src/main/java/com/example/spinlog/statistics/service/wordanalysis/MorphemeExtractionService.java
@@ -1,4 +1,4 @@
-package com.example.spinlog.statistics.service.workanalysis;
+package com.example.spinlog.statistics.service.wordanalysis;
 
 import com.example.spinlog.statistics.service.dto.MorphemeApiRequestDto;
 import com.example.spinlog.statistics.service.dto.MorphemeResponse;
